requirements.txt
streamlit>=1.29.0
streamlit-authenticator
streamlit-webrtc>=0.45.1
opencv-python>=4.10.0.84
numpy>=1.26.4
redis>=5.2.1
pandas>=2.2.3
PyYAML>=6.0.2
av>=13.1
bcrypt>=4.2.1 
pillow>=10.4.0
scikit-learn>=1.5.2
insightface
onnxruntime